refactor(accordion-web): move accordion groups reducer
