+++ packages/pluggableWidgets/accordion-web/src/components/Accordion.tsx
@@ -1,7 +1,7 @@
 import { createContext, createElement, Dispatch, ReactElement, useMemo, useReducer, useRef } from "react";
 
 import AccordionGroup, { AccGroup } from "./AccordionGroup";
-
+import { AccordionGroupsReducerAction, getAccordionGroupsReducer } from "../utils/AccordionGroupStateReducer";
 import { AccordionContainerProps } from "../../typings/AccordionProps";
 
 interface AccordionProps extends Pick<AccordionContainerProps, "class" | "style" | "tabIndex"> {
@@ -9,46 +9,6 @@ interface AccordionProps extends Pick<AccordionContainerProps, "class" | "style"
     groups: AccGroup[];
     collapsible: boolean;
     singleExpandedGroup?: boolean;
-}
-
-type AccordionGroupsReducerAction =
-    | { type: "sync"; groups: AccGroup[] }
-    | { type: "expand" | "collapse"; group: AccGroup };
-
-export function getAccordionGroupsReducer(
-    expand: "single" | "multiple" | "all"
-): (state: AccGroup[], action: AccordionGroupsReducerAction) => AccGroup[] {
-    return (state: AccGroup[], action: AccordionGroupsReducerAction): AccGroup[] => {
-        if (action.type === "sync" || expand === "single") {
-            const newState = action.type === "sync" ? action.groups : state;
-
-            return newState.map((group, index) => {
-                let collapsed;
-
-                if (action.type === "sync") {
-                    collapsed = state[index].collapsed; // The previous collapsed state of the group at the same index can be used, because the order of groups remains the same throughout the lifetime of the widget.
-                } else {
-                    collapsed = action.type === "expand" ? group !== action.group : true;
-                }
-
-                return {
-                    ...group,
-                    collapsed
-                };
-            });
-        }
-
-        const newState = [...state];
-        const groupIndex = newState.findIndex(group => group === action.group);
-
-        if (action.type === "expand") {
-            newState[groupIndex] = { ...action.group, collapsed: false };
-            return newState;
-        }
-
-        newState[groupIndex] = { ...action.group, collapsed: true };
-        return newState;
-    };
 }
 
 export const AccordionGroupsDispatch = createContext<Dispatch<AccordionGroupsReducerAction> | undefined>(undefined);
